Prevent reloading main menu if null file chosen
Whenever someone loads a new profile, the main menu is reloaded
to update the view for that profile. When the main menu is
reloaded, the notifications tray opens. If someone did not
actually choose a new profile, the only discernable change to the
view is seeing the notifications tray open. It is best to not
reload the main menu to prevent the notifications tray from being
opened.